add library wrapper to encapsulate scrollview and tableview
