+++ source/views/settings/screens/overview/component-library/context-menu.tsx
@@ -1,36 +1,35 @@
 import * as React from 'react'
 import {ContextMenu} from '@frogpond/context-menu'
-import {TableView, Section, Cell} from '@frogpond/tableview'
+import {Cell, Section} from '@frogpond/tableview'
 import {Button} from '@frogpond/button'
+import {LibraryWrapper} from './base/library-wrapper'
 import {upperFirst} from 'lodash'
 
 const SingleMenu = (): JSX.Element => {
 	const [value, setValue] = React.useState('dog')
 
 	return (
-		<Cell
-			cellAccessoryView={
-				<ContextMenu
-					actions={['bird', 'cat', 'cow', 'dog']}
-					isMenuPrimaryAction={true}
-					onPressMenuItem={setValue}
-					title="Select an animal."
-				>
-					<Button title={upperFirst(value)} />
-				</ContextMenu>
-			}
-			cellStyle="RightDetail"
-			title="Top-level menu"
-		/>
+		<Section>
+			<Cell
+				cellAccessoryView={
+					<ContextMenu
+						actions={['bird', 'cat', 'cow', 'dog']}
+						isMenuPrimaryAction={true}
+						onPressMenuItem={setValue}
+						title="Select an animal."
+					>
+						<Button title={upperFirst(value)} />
+					</ContextMenu>
+				}
+				cellStyle="RightDetail"
+				title="Top-level menu"
+			/>
+		</Section>
 	)
 }
 
-export const ContextMenuLibrary = (): JSX.Element => {
-	return (
-		<TableView>
-			<Section>
-				<SingleMenu />
-			</Section>
-		</TableView>
-	)
-}
+export const ContextMenuLibrary = (): JSX.Element => (
+	<LibraryWrapper>
+		<SingleMenu />
+	</LibraryWrapper>
+)
